correspond to js and creating popup form
