+++ app/javascript/packs/application.js
@@ -8,7 +8,6 @@ import Turbolinks from "turbolinks"
 import * as ActiveStorage from "@rails/activestorage"
 import "channels"
 
-// require("../stylesheets/application.css")
 require("../stylesheets/tailwindcss.css")
 import "stylesheets/application"
 
@@ -16,4 +15,5 @@ Rails.start()
 Turbolinks.start()
 ActiveStorage.start()
 
-import "stylesheets/application"
+// require("../UI/index.js")
+import "./UI/index"
