--- app/javascript/packs/service/index.js
@@ -1,27 +1,0 @@
-import { inputChangeSelector } from "@rails/ujs";
-import { ContextExclusionPlugin } from "webpack";
-
-function popupImage() {
-    var popup = document.getElementById('js-popup');
-    if (!popup) return;
-
-    var blackBg = document.getElementById('js-black-bg');
-    var closeBtn = document.getElementById('js-close-btn');
-    var showBtn = document.getElementById('js-show-popup');
-
-    closePopUp(blackBg);
-    closePopUp(closeBtn);
-    closePopUp(showBtn);
-    function closePopUp(elem) {
-        if (!elem) return;
-        elem.addEventListener('click', function () {
-            popup.classList.toggle('is-show');
-        });
-    }
-}
-popupImage();
-
-window.onload = function () {
-    console.log("loaded");
-    alert("hoge");
-}
